+++ core/item.go
@@ -175,8 +175,8 @@ func (k *KeyPair) WriteItem(obj *Item) error {
 }
 
 // SaveActiveItems saves all active items from rooms and characters to the database.
-func (s *Server) SaveActiveItems() error {
-	if s == nil {
+func (g *Game) SaveActiveItems() error {
+	if g == nil {
 		return fmt.Errorf("server is nil")
 	}
 
@@ -186,8 +186,8 @@ func (s *Server) SaveActiveItems() error {
 	itemsToSave := make(map[uuid.UUID]*Item)
 
 	// Items in rooms
-	if s.Rooms != nil {
-		for roomID, room := range s.Rooms {
+	if g.Rooms != nil {
+		for roomID, room := range g.Rooms {
 			if room == nil {
 				Logger.Warn("Nil room found", "roomID", roomID)
 				continue
@@ -207,8 +207,8 @@ func (s *Server) SaveActiveItems() error {
 	}
 
 	// Items in character inventories
-	if s.Characters != nil {
-		for charID, character := range s.Characters {
+	if g.Characters != nil {
+		for charID, character := range g.Characters {
 			if character == nil {
 				Logger.Warn("Nil character found", "characterID", charID)
 				continue
@@ -228,7 +228,7 @@ func (s *Server) SaveActiveItems() error {
 	}
 
 	// Save all collected items
-	if s.Database == nil {
+	if g.Database == nil {
 		return fmt.Errorf("server database is nil")
 	}
 
@@ -245,7 +245,7 @@ func (s *Server) SaveActiveItems() error {
 		}
 
 		// Attempt to write the item to the database
-		if err := s.Database.WriteItem(item); err != nil {
+		if err := g.Database.WriteItem(item); err != nil {
 			Logger.Error("Error saving item", "itemName", item.Name, "itemID", item.ID, "error", err)
 			// Continue saving other items even if one fails
 		} else {
@@ -259,8 +259,8 @@ func (s *Server) SaveActiveItems() error {
 	return nil
 }
 
-func (s *Server) CreateItemFromPrototype(prototypeID uuid.UUID) (*Item, error) {
-	prototype, exists := s.Prototypes[prototypeID]
+func (g *Game) CreateItemFromPrototype(prototypeID uuid.UUID) (*Item, error) {
+	prototype, exists := g.Prototypes[prototypeID]
 	if !exists {
 		Logger.Error("Prototype not found", "prototypeID", prototypeID)
 		return nil, fmt.Errorf("prototype with ID %s not found", prototypeID)
@@ -302,7 +302,7 @@ func (s *Server) CreateItemFromPrototype(prototypeID uuid.UUID) (*Item, error) {
 	if newItem.Container {
 		newItem.Contents = make([]*Item, 0, len(prototype.Contents))
 		for _, contentProtoID := range prototype.Contents {
-			newContentItem, err := s.CreateItemFromPrototype(contentProtoID)
+			newContentItem, err := g.CreateItemFromPrototype(contentProtoID)
 			if err != nil {
 				Logger.Error("Error creating content item from prototype", "prototypeID", contentProtoID, "error", err)
 				continue // Skip this content item but continue with others
@@ -312,7 +312,7 @@ func (s *Server) CreateItemFromPrototype(prototypeID uuid.UUID) (*Item, error) {
 	}
 
 	// Save the new item to the database
-	if err := s.Database.WriteItem(newItem); err != nil {
+	if err := g.Database.WriteItem(newItem); err != nil {
 		Logger.Error("Failed to write new item to database", "itemName", newItem.Name, "itemID", newItem.ID, "error", err)
 		return nil, fmt.Errorf("failed to write new item to database: %w", err)
 	}
